requirements.txt
click==8.0.3
flake8==4.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
python-dotenv==0.19.2
